Refactor module structure and add documentation
- Split implementation out of mod.rs files into dedicated modules.
- Update contributing/ARCHITECTURE.md to match new module structure
- Add comprehensive module-level docs.
- Minor README improvements.
