+++ Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "mom-rpc"
-version = "0.5.2"
+version = "0.5.3"
 edition = "2021"
 authors = ["John Basrai"]
 description = "Transport-agnostic async RPC over message-oriented middleware (AMQP, MQTT, in-memory)"
